+++ src/UI/Fields/GasLimit/tests/GasLimit.test.ts
@@ -1,109 +1,13 @@
-import { GAS_LIMIT } from '@multiversx/sdk-dapp/constants/index';
 import { fireEvent, waitFor } from '@testing-library/react';
 import { testAddress } from '__mocks__';
 import { formattedAmountSelector } from 'tests/helpers';
 import { renderForm } from 'tests/helpers/renderForm';
-import { ValuesEnum } from 'types';
 
 describe('GasLimit field', () => {
   it('should not be empty', async () => {
     const methods = renderForm();
     const input: any = await methods.findByLabelText('Gas Limit');
     expect(input.value).toBe('50000');
-  });
-  it('should not be string', async () => {
-    const { findByLabelText, findByText } = await renderForm();
-    const input: any = await findByLabelText('Gas Limit');
-    const value = 'string';
-    const data = { target: { value } };
-    fireEvent.change(input, data);
-    fireEvent.blur(input);
-    const req = await findByText('Invalid number');
-    expect(req?.innerHTML).toBe('Invalid number');
-  });
-  it('should be integer', async () => {
-    const { findByLabelText, queryByText } = await renderForm();
-    const input: any = await findByLabelText('Gas Limit');
-    const value = '0.1';
-    const data = { target: { value } };
-    fireEvent.change(input, data);
-    fireEvent.blur(input);
-    await waitFor(() => {
-      const req = queryByText('Invalid number');
-      expect(req?.innerHTML).toBe('Invalid number');
-    });
-  });
-  it('should not allow exponential gasLimit', async () => {
-    const { findByLabelText, queryByText } = await renderForm();
-    const input: any = await findByLabelText('Gas Limit');
-    const value = '1e20';
-    const data = { target: { value } };
-    fireEvent.change(input, data);
-    fireEvent.blur(input);
-    await waitFor(() => {
-      const req = queryByText('Invalid number');
-      expect(req?.innerHTML).toBe('Invalid number');
-    });
-  });
-  it('should not allow negative gasLimit', async () => {
-    const { findByLabelText, queryByText } = await renderForm();
-    const input: any = await findByLabelText('Gas Limit');
-    const value = '-1';
-    const data = { target: { value } };
-    fireEvent.change(input, data);
-    fireEvent.blur(input);
-    await waitFor(() => {
-      const req = queryByText('Invalid number');
-      expect(req?.innerHTML).toBeDefined();
-    });
-  });
-  it('should not allow explicit positive gasLimit', async () => {
-    const { findByLabelText, queryByText } = await renderForm();
-    const input: any = await findByLabelText('Gas Limit');
-    const value = '+1';
-    const data = { target: { value } };
-    fireEvent.change(input, data);
-    fireEvent.blur(input);
-    await waitFor(() => {
-      const req = queryByText('Invalid number');
-      expect(req?.innerHTML).toBeDefined();
-    });
-  });
-  it('should >= than the one set by config', async () => {
-    const { findByLabelText, queryByText } = await renderForm();
-    const input: any = await findByLabelText('Gas Limit');
-    const value = GAS_LIMIT - 1;
-    const data = { target: { value } };
-    fireEvent.change(input, data);
-    fireEvent.blur(input);
-
-    await waitFor(() => {
-      const req = queryByText(/^Gas limit must be greater/);
-      expect(req?.innerHTML).toBe(
-        `Gas limit must be greater or equal to ${GAS_LIMIT}`
-      );
-    });
-  });
-  it('should >= than the configGasLimit + data.length if data is set', async () => {
-    const methods = renderForm();
-
-    const dataInput: any = await methods.findByTestId(ValuesEnum.data);
-    const dataValue = 'four';
-    fireEvent.change(dataInput, { target: { value: dataValue } });
-    fireEvent.blur(dataInput);
-
-    const input: any = methods.getByLabelText('Gas Limit');
-    const value = GAS_LIMIT;
-    const data = { target: { value } };
-    fireEvent.change(input, data);
-    fireEvent.blur(input);
-
-    await waitFor(() => {
-      const req = methods.queryByText(/^Gas limit must be greater/);
-      expect(req?.innerHTML).toBe(
-        'Gas limit must be greater or equal to 56000'
-      );
-    });
   });
   it('setting Gas limit + amount > balance should trigger error', async () => {
     const { getByLabelText, getByTestId, queryByText, findByTestId } =