+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Changed `nonce` type from `string` to `number`](https://github.com/multiversx/mx-sdk-dapp-form/pull/173)
 
 ## [[0.6.0](https://github.com/multiversx/mx-sdk-dapp-form/pull/172)] - 2023-05-09
 - [Verify ledger version for guarded transactions](https://github.com/multiversx/mx-sdk-dapp-form/pull/171)